feat(object_oriented): Update configuration to use object oriented design
